Switch to GDI+ to allow window background toggling
Regular GDI was proving difficult to have a visual keypress rectangle
shown while the rest of window invisible.  Furthermore, there was no
obvious way of fading out the keypress rectangle over a short time
period when the user hasn't pressed any more keys within that time.

Given how the plan is to fade both the rounded rectangle and the
keypress text I began looking for alternative methods of providing
such functionality.  The two choices boiled down to GDI+ and
DirectDraw.  Despite being a software rendered approach, GDI+ appeared
to be the simpler of the two API's, provided the exact functionality
needed, and is an older/longer supported API.  Given how this
application isn't supposed to be graphically intensive, the fact that
GDI+ uses a software rendered approach shouldn't be an issue.