Detect if shift modifier affects key press
get_key has been replaced with get_key_info which returns a KeyInfo
object that allows the caller to determine if the shift key affects
its display representation.  For example, the 'j' key would become the
'J' key and the KeyInfo object would state that the shift key does
affect this representation.  On the other hand, the 'F2' key would
state that the shift modifier has no affect on this key press.