Always return DefWindowProc after a WM_PAINT message
By returning 0 after rendering the application from a WM_PAINT message
while the window was in layered/hidden mode would cause high CPU usage
as the application was getting flooded with WM_PAINT calls.  Omitting
the statement and allowing DefWindowProc to be called after rendering
was complete alleviated this and reduced usage to 0.5% when keys were
actually being rendered.  Also, no other issues were observed when the
window was not in layered/hidden mode.